backend/requirements.txt
torch==2.11.0
torchvision==0.26.0
fastapi==0.136.0
uvicorn==0.44.0
python-multipart==0.0.26
Pillow==12.2.0
numpy==2.4.4
